@@ -18,7 +18,7 @@
 #ifndef HEADER_SUPERTUX_CONTROL_KEYBOARD_MANAGER_HPP
 #define HEADER_SUPERTUX_CONTROL_KEYBOARD_MANAGER_HPP
 
-#include <boost/optional.hpp>
+#include <optional>
 
 #include "control/controller.hpp"
 
@@ -42,7 +42,7 @@ public:
 private:
   InputManager* m_parent;
   KeyboardConfig& m_keyboard_config;
-  boost::optional<Control> m_wait_for_key;
+  std::optional<Control> m_wait_for_key;
   bool m_lock_text_input;
 
 private:
