@@ -289,7 +289,7 @@ Editor::get_level_directory() const
 }
 
 void
-Editor::test_level(const boost::optional<std::pair<std::string, Vector>>& test_pos)
+Editor::test_level(const std::optional<std::pair<std::string, Vector>>& test_pos)
 {
 
   Tile::draw_editor_images = false;
@@ -649,7 +649,7 @@ Editor::event(const SDL_Event& ev)
 	if (ev.type == SDL_KEYDOWN &&
         ev.key.keysym.sym == SDLK_t &&
         ev.key.keysym.mod & KMOD_CTRL) {
-		test_level(boost::none);
+		test_level(std::nullopt);
 		}
 
 	if (ev.type == SDL_KEYDOWN &&
