@@ -17,7 +17,7 @@
 #ifndef HEADER_SUPERTUX_SUPERTUX_GAME_MANAGER_HPP
 #define HEADER_SUPERTUX_SUPERTUX_GAME_MANAGER_HPP
 
-#include <boost/optional.hpp>
+#include <optional>
 #include <memory>
 #include <string>
 #include "math/vector.hpp"
@@ -33,7 +33,7 @@ public:
 
   void start_worldmap(const World& world, const std::string& spawnpoint = "", const std::string& worldmap_filename = "");
   void start_level(const World& world, const std::string& level_filename,
-                   const boost::optional<std::pair<std::string, Vector>>& start_pos = boost::none);
+                   const std::optional<std::pair<std::string, Vector>>& start_pos = std::nullopt);
 
   bool load_next_worldmap();
   void set_next_worldmap(const std::string& worldmap, const std::string &spawnpoint);
