@@ -19,11 +19,10 @@
 #include <config.h>
 #include <version.h>
 #include <fstream>
+#include <filesystem>
 
 #include <SDL_image.h>
 #include <SDL_ttf.h>
-#include <boost/filesystem.hpp>
-#include <boost/locale.hpp>
 #include <physfs.h>
 #include <tinygettext/log.hpp>
 extern "C" {
@@ -155,13 +154,13 @@ Main::init_tinygettext()
 class PhysfsSubsystem final
 {
 private:
-  boost::optional<std::string> m_forced_datadir;
-  boost::optional<std::string> m_forced_userdir;
+  std::optional<std::string> m_forced_datadir;
+  std::optional<std::string> m_forced_userdir;
 
 public:
   PhysfsSubsystem(const char* argv0,
-                  boost::optional<std::string> forced_datadir,
-                  boost::optional<std::string> forced_userdir) :
+                  std::optional<std::string> forced_datadir,
+                  std::optional<std::string> forced_userdir) :
     m_forced_datadir(std::move(forced_datadir)),
     m_forced_userdir(std::move(forced_userdir))
   {
@@ -203,7 +202,7 @@ public:
       {
         datadir = BUILD_DATA_DIR;
         // Add config dir for supplemental files
-        PHYSFS_mount(boost::filesystem::canonical(BUILD_CONFIG_DATA_DIR).string().c_str(), nullptr, 1);
+        PHYSFS_mount(std::filesystem::canonical(BUILD_CONFIG_DATA_DIR).string().c_str(), nullptr, 1);
       }
       else
       {
@@ -214,7 +213,7 @@ public:
       }
     }
 
-    if (!PHYSFS_mount(boost::filesystem::canonical(datadir).string().c_str(), nullptr, 1))
+    if (!PHYSFS_mount(std::filesystem::canonical(datadir).string().c_str(), nullptr, 1))
     {
       log_warning << "Couldn't add '" << datadir << "' to physfs searchpath: " << PHYSFS_getLastErrorCode() << std::endl;
     }
@@ -252,20 +251,20 @@ public:
 	std::string olduserdir = FileSystem::join(physfs_userdir, "." PACKAGE_NAME);
 #endif
 	if (FileSystem::is_directory(olduserdir)) {
-	  boost::filesystem::path olduserpath(olduserdir);
-	  boost::filesystem::path userpath(userdir);
+	  std::filesystem::path olduserpath = std::filesystem::u8path(olduserdir);
+	  std::filesystem::path userpath = std::filesystem::u8path(userdir);
 
-	  boost::filesystem::directory_iterator end_itr;
+	  std::filesystem::directory_iterator end_itr;
 
 	  bool success = true;
 
 	  // cycle through the directory
-	  for (boost::filesystem::directory_iterator itr(olduserpath); itr != end_itr; ++itr) {
+	  for (std::filesystem::directory_iterator itr(olduserpath); itr != end_itr; ++itr) {
 		try
 		{
-		  boost::filesystem::rename(itr->path().string().c_str(), userpath / itr->path().filename());
+		  std::filesystem::rename(itr->path().string().c_str(), userpath / itr->path().filename());
 		}
-		catch (const boost::filesystem::filesystem_error& err)
+		catch (const std::filesystem::filesystem_error& err)
 		{
 		  success = false;
 		  log_warning << "Failed to move contents of config directory: " << err.what() << std::endl;
@@ -274,9 +273,9 @@ public:
 	  if (success) {
 	    try
 		{
-		  boost::filesystem::remove_all(olduserpath);
+		  std::filesystem::remove_all(olduserpath);
 		}
-		catch (const boost::filesystem::filesystem_error& err)
+		catch (const std::filesystem::filesystem_error& err)
 		{
 		  success = false;
 		  log_warning << "Failed to remove old config directory: " << err.what();
@@ -505,12 +504,12 @@ Main::launch_game(const CommandLineArguments& args)
 
         if (args.sector || args.spawnpoint)
         {
-          std::string sectorname = args.sector.get_value_or("main");
+          std::string sectorname = args.sector.value_or("main");
 
           const auto& spawnpoints = session->get_current_sector().get_objects_by_type<SpawnPointMarker>();
           std::string default_spawnpoint = (spawnpoints.begin() != spawnpoints.end()) ?
             "" : spawnpoints.begin()->get_name();
-          std::string spawnpointname = args.spawnpoint.get_value_or(default_spawnpoint);
+          std::string spawnpointname = args.spawnpoint.value_or(default_spawnpoint);
 
           session->set_start_point(sectorname, spawnpointname);
           session->restart_level();
