@@ -264,7 +264,7 @@ SectorParser::parse_old_format(const ReaderMapping& reader)
   }
 
   // read reset-points (now spawn-points)
-  boost::optional<ReaderMapping> resetpoints;
+  std::optional<ReaderMapping> resetpoints;
   if (reader.get("reset-points", resetpoints)) {
     auto iter = resetpoints->get_iter();
     while (iter.next()) {
@@ -281,7 +281,7 @@ SectorParser::parse_old_format(const ReaderMapping& reader)
   }
 
   // read objects
-  boost::optional<ReaderCollection> objects;
+  std::optional<ReaderCollection> objects;
   if (reader.get("objects", objects)) {
     for (auto const& obj : objects->get_objects())
     {
