@@ -161,13 +161,13 @@ TileSetParser::parse_tile(const ReaderMapping& reader)
   }
 
   std::vector<SurfacePtr> editor_surfaces;
-  boost::optional<ReaderMapping> editor_images_mapping;
+  std::optional<ReaderMapping> editor_images_mapping;
   if (reader.get("editor-images", editor_images_mapping)) {
     editor_surfaces = parse_imagespecs(*editor_images_mapping);
   }
 
   std::vector<SurfacePtr> surfaces;
-  boost::optional<ReaderMapping> images_mapping;
+  std::optional<ReaderMapping> images_mapping;
   if (reader.get("images", images_mapping)) {
     surfaces = parse_imagespecs(*images_mapping);
   }
@@ -257,13 +257,13 @@ TileSetParser::parse_tiles(const ReaderMapping& reader)
     if (shared_surface)
     {
       std::vector<SurfacePtr> editor_surfaces;
-      boost::optional<ReaderMapping> editor_surfaces_mapping;
+      std::optional<ReaderMapping> editor_surfaces_mapping;
       if (reader.get("editor-images", editor_surfaces_mapping)) {
         editor_surfaces = parse_imagespecs(*editor_surfaces_mapping);
       }
 
       std::vector<SurfacePtr> surfaces;
-      boost::optional<ReaderMapping> surfaces_mapping;
+      std::optional<ReaderMapping> surfaces_mapping;
       if (reader.get("image", surfaces_mapping) ||
          reader.get("images", surfaces_mapping)) {
         surfaces = parse_imagespecs(*surfaces_mapping);
@@ -310,14 +310,14 @@ TileSetParser::parse_tiles(const ReaderMapping& reader)
           int y = static_cast<int>(32 * (i / width));
 
           std::vector<SurfacePtr> surfaces;
-          boost::optional<ReaderMapping> surfaces_mapping;
+          std::optional<ReaderMapping> surfaces_mapping;
           if (reader.get("image", surfaces_mapping) ||
              reader.get("images", surfaces_mapping)) {
             surfaces = parse_imagespecs(*surfaces_mapping, Rect(x, y, Size(32, 32)));
           }
 
           std::vector<SurfacePtr> editor_surfaces;
-          boost::optional<ReaderMapping> editor_surfaces_mapping;
+          std::optional<ReaderMapping> editor_surfaces_mapping;
           if (reader.get("editor-images", editor_surfaces_mapping)) {
             editor_surfaces = parse_imagespecs(*editor_surfaces_mapping, Rect(x, y, Size(32, 32)));
           }
@@ -337,7 +337,7 @@ TileSetParser::parse_tiles(const ReaderMapping& reader)
 
 std::vector<SurfacePtr>
   TileSetParser::parse_imagespecs(const ReaderMapping& images_mapping,
-                                  const boost::optional<Rect>& surface_region) const
+                                  const std::optional<Rect>& surface_region) const
 {
   std::vector<SurfacePtr> surfaces;
 
